+++ src/crawlers/crawler_utils.js
@@ -1,6 +1,8 @@
+import { Cookie, CookieJar } from 'tough-cookie';
 // TYPE IMPORTS
 /* eslint-disable no-unused-vars,import/named,import/no-duplicates,import/order */
 import { Session } from '../session_pool/session';
+import log from '../utils_log';
 /* eslint-enable no-unused-vars,import/named,import/no-duplicates,import/order */
 
 /**
@@ -28,4 +30,61 @@ export function throwOnBlockedRequest(session, statusCode) {
     if (isBlocked) {
         throw new Error(`Request blocked - received ${statusCode} status code.`);
     }
+}
+
+/**
+ * Merges multiple cookie strings. Keys are compared case-sensitively, warning will be logged
+ * if we see two cookies with same keys but different casing.
+ *
+ * @param {string} url
+ * @param {string[]} sourceCookies
+ * @return {string}
+ * @private
+ */
+export function mergeCookies(url, sourceCookies) {
+    const jar = new CookieJar();
+
+    // ignore empty cookies
+    for (const sourceCookieString of sourceCookies.filter((c) => c)) {
+        const cookies = sourceCookieString.split(/ *; */); // ignore extra spaces
+
+        for (const cookieString of cookies) {
+            const cookie = Cookie.parse(cookieString);
+            const similarKeyCookie = jar.getCookiesSync(url).find((c) => {
+                return cookie.key !== c.key && cookie.key.toLowerCase() === c.key.toLowerCase();
+            });
+
+            if (similarKeyCookie) {
+                log.deprecated(`Found cookies with similar name during cookie merging: '${cookie.key}' and '${similarKeyCookie.key}'`);
+            }
+
+            jar.setCookieSync(cookie, url);
+        }
+    }
+
+    return jar.getCookieStringSync(url);
+}
+
+/**
+ * @param {string} url
+ * @param {string} cookieString1
+ * @param {string} cookieString2
+ * @return {string}
+ * @private
+ */
+export function diffCookies(url, cookieString1, cookieString2) {
+    if (cookieString1 === cookieString2) {
+        return '';
+    }
+
+    const cookies1 = cookieString1.split(/ *; */).map((cookie) => Cookie.parse(cookie));
+    const cookies2 = cookieString2.split(/ *; */).map((cookie) => Cookie.parse(cookie));
+
+    const added = cookies2.filter((newCookie) => {
+        return !cookies1.find((oldCookie) => newCookie.toString() === oldCookie.toString());
+    });
+    const jar = new CookieJar();
+    added.forEach((cookie) => jar.setCookieSync(cookie, url));
+
+    return jar.getCookieStringSync(url);
 }