+++ test/crawlers/cheerio_crawler.test.js
@@ -8,6 +8,7 @@ import bodyParser from 'body-parser';
 import sinon from 'sinon';
 import { Readable } from 'stream';
 import iconv from 'iconv-lite';
+import { Log } from '@apify/log';
 import log from '../../build/utils_log';
 import Apify from '../../build';
 import { sleep } from '../../build/utils';
@@ -18,6 +19,7 @@ import * as utilsRequest from '../../build/utils_request';
 import CrawlerExtension from '../../build/crawlers/crawler_extension';
 import Request from '../../build/request';
 import AutoscaledPool from '../../build/autoscaling/autoscaled_pool';
+import { mergeCookies } from '../../build/crawlers/crawler_utils';
 
 const HOST = '127.0.0.1';
 
@@ -94,6 +96,10 @@ app.post('/mock', (req, res) => {
     Object.entries(headers).forEach(([key, value]) => res.setHeader(key, value));
 
     res.status(statusCode).send(body);
+});
+
+app.get('/headers', (req, res) => {
+    res.status(200).json({ headers: req.headers });
 });
 
 app.get('/invalidContentType', (req, res) => {
@@ -904,12 +910,113 @@ describe('CheerioCrawler', () => {
 
             });
 
+            const spy = jest.spyOn(utilsRequest, 'requestAsBrowser');
             await crawler.run();
             requests.forEach((req, i) => {
                 if (i >= 1) {
-                    expect(req.headers.Cookie).toEqual(cookie);
+                    expect(spy.mock.calls[i][0].headers.Cookie).toBe(cookie);
                 }
             });
+            spy.mockRestore();
+        });
+
+        test('should merge cookies set in pre-nav hook with the session ones', async () => {
+            const responses = [];
+            const requestAsBrowserOptions = [];
+            const crawler = new Apify.CheerioCrawler({
+                requestList: await Apify.openRequestList(null, [{
+                    url: `http://${HOST}:${port}/headers`,
+                    headers: { cookie: 'foo=bar2; baz=123' },
+                }]),
+                useSessionPool: true,
+                persistCookiesPerSession: false,
+                sessionPoolOptions: {
+                    maxPoolSize: 1,
+                },
+                handlePageFunction: async ({ json }) => {
+                    responses.push(json);
+                },
+                preNavigationHooks: [(context, options) => {
+                    requestAsBrowserOptions.push(options);
+                }],
+            });
+
+            const sessSpy = jest.spyOn(Session.prototype, 'getCookieString');
+            sessSpy.mockReturnValueOnce('foo=bar1; other=cookie1; coo=kie');
+            await crawler.run();
+            expect(responses).toHaveLength(1);
+            expect(responses[0]).toMatchObject({
+                headers: {
+                    cookie: 'foo=bar2; other=cookie1; coo=kie; baz=123',
+                },
+            });
+            expect(requestAsBrowserOptions).toHaveLength(1);
+            expect(requestAsBrowserOptions[0]).toMatchObject({
+                headers: {
+                    Cookie: 'foo=bar2; other=cookie1; coo=kie; baz=123', // header name normalized to `Cookie`
+                },
+            });
+        });
+
+        test('should work with cookies adjusted on `context.request` in pre-nav hook', async () => {
+            const responses = [];
+            const crawler = new Apify.CheerioCrawler({
+                requestList: await Apify.openRequestList(null, [{
+                    url: `http://${HOST}:${port}/headers`,
+                    headers: { cookie: 'foo=bar2; baz=123' },
+                }]),
+                useSessionPool: true,
+                persistCookiesPerSession: false,
+                sessionPoolOptions: {
+                    maxPoolSize: 1,
+                },
+                handlePageFunction: async ({ json }) => {
+                    responses.push(json);
+                },
+                preNavigationHooks: [({ request }) => {
+                    request.headers.Cookie = 'foo=override; coo=kie';
+                }],
+            });
+
+            await crawler.run();
+            expect(responses).toHaveLength(1);
+            expect(responses[0]).toMatchObject({
+                headers: {
+                    cookie: 'foo=override; baz=123; coo=kie',
+                },
+            });
+        });
+
+        test('mergeCookies()', async () => {
+            const deprecatedSpy = jest.spyOn(Log.prototype, 'deprecated');
+            const cookie1 = mergeCookies('https://example.com', [
+                'foo=bar1; other=cookie1 ; coo=kie',
+                'foo=bar2; baz=123',
+                'other=cookie2;foo=bar3',
+            ]);
+            expect(cookie1).toBe('foo=bar3; other=cookie2; coo=kie; baz=123');
+            expect(deprecatedSpy).not.toBeCalled();
+
+            const cookie2 = mergeCookies('https://example.com', [
+                'Foo=bar1; other=cookie1 ; coo=kie',
+                'foo=bar2; baz=123',
+                'Other=cookie2;foo=bar3',
+            ]);
+            expect(cookie2).toBe('Foo=bar1; other=cookie1; coo=kie; foo=bar3; baz=123; Other=cookie2');
+            expect(deprecatedSpy).toBeCalledTimes(3);
+            expect(deprecatedSpy).toBeCalledWith(`Found cookies with similar name during cookie merging: 'foo' and 'Foo'`);
+            expect(deprecatedSpy).toBeCalledWith(`Found cookies with similar name during cookie merging: 'Other' and 'other'`);
+            deprecatedSpy.mockClear();
+
+            const cookie3 = mergeCookies('https://example.com', [
+                'foo=bar1; Other=cookie1 ; Coo=kie',
+                'foo=bar2; baz=123',
+                'Other=cookie2;Foo=bar3;coo=kee',
+            ]);
+            expect(cookie3).toBe('foo=bar2; Other=cookie2; Coo=kie; baz=123; Foo=bar3; coo=kee');
+            expect(deprecatedSpy).toBeCalledTimes(2);
+            expect(deprecatedSpy).toBeCalledWith(`Found cookies with similar name during cookie merging: 'Foo' and 'foo'`);
+            expect(deprecatedSpy).toBeCalledWith(`Found cookies with similar name during cookie merging: 'coo' and 'Coo'`);
         });
 
         test('should pass session to prepareRequestFunction when Session pool is used', async () => {