+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - support channel and user links in YouTube regex (#1178)
 - fix: allow passing `requestsFromUrl` to `RequestListOptions` in TS (#1191)
 - allow passing `forceCloud` down to the KV store (#1186), closes #752
+- merge cookies from session with user provided ones (#1201), closes #1197
 
 2.0.7 / 2021/09/08
 ====================